Detect Mac OS X systems and adapt accordingly
Mac OS X does not have a poweroff utility, therefore exclude it from
installation on this system. Also, there is no group named root and
executable helpers should go into libexec in the selected prefix
directory.

Basing this decision on uname may not be correct as that is more about
the kernel, but it is good enough at the moment.